fix somebugs

+++ lib/searpc-server.c
@@ -134,7 +134,7 @@ searpc_set_objlist_to_ret_object (json_t *object, GList *ret)
 }
 
 char *
-searpc_marshal_set_ret_common (json_t *object, int *len,  GError *error)
+searpc_marshal_set_ret_common (json_t *object, gsize *len,  GError *error)
 {
 
     char *data;
